[DDO-32] Extract LogFile/DnsRecord classes to own modules
The LogFile class is used outside of the LogManager class. Move it to
its own module

The DnsRecord class is used outside of the EdgeDnsManager class. Move it
to its own module
