+++ lds_connector/config.py
@@ -15,10 +15,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from dataclasses import dataclass
 import logging
 import os
+from dataclasses import dataclass
 from typing import Optional
+
 import yaml
