+++ lds_connector/dns_record.py
@@ -1,0 +1,28 @@
+# Original author: Cam Mackintosh <cmackint@akamai.com>
+# For more information visit https://developer.akamai.com
+
+# Copyright 2023 Akamai Technologies, Inc. All Rights Reserved
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from dataclasses import dataclass
+
+
+@dataclass
+class DnsRecord:
+    time_fetched_sec: float
+    zone: str
+    name: str
+    type: str
+    ttl_sec: int
+    rdata: list[str]
